docs: add version tip
